Move build handler to async worker
