+++ Procfile
@@ -1,1 +1,2 @@
 web: gunicorn linty.wsgi --log-file - --reload
+worker: python manage.py rqworker
