+++ interface/views.py
@@ -1,11 +1,7 @@
 import hashlib
 import hmac
 import json
-import os
-import shutil
-import subprocess
 
-import requests
 from django.conf import settings
 from django.contrib.auth import logout
 from django.contrib.auth.decorators import login_required
@@ -14,7 +10,6 @@ from django.core.urlresolvers import reverse
 from django.db.models import Count
 from django.http import HttpResponse, Http404
 from django.shortcuts import redirect, render
-from django.utils import timezone
 from django.utils.decorators import method_decorator
 from django.views import generic
 from django.views.decorators.csrf import csrf_exempt
@@ -23,6 +18,7 @@ from social.apps.django_app.default.models import UserSocialAuth
 from social.apps.django_app.views import auth
 
 from interface.models import Build, Repo
+from interface.tasks import build_handler
 from interface.utils import get_github
 
 
@@ -177,84 +173,10 @@ def WebhookView(request):
     if 'ref' not in body or not body['head_commit']:
         return HttpResponse(status=204)
 
-    try:
-        repo = Repo.objects.get(full_name=body['repository']['full_name'])
-    except Repo.DoesNotExist:
-        return HttpResponse(status=204)
+    base_url = request.build_absolute_uri('/')[:-1]
+    build_handler.delay(body, base_url)
 
-    try:
-        data = UserSocialAuth.objects.filter(user=repo.user).values_list('extra_data')[0][0]
-    except:
-        raise Exception('Fail')
-
-    username = data['login']
-    password = data['access_token']
-    auth = (username, password)
-
-    # get necessary vars
-    clone_url = body['repository']['clone_url']
-    clone_url = clone_url.replace('github.com', '%s:%s@github.com' % (username, password))
-    branch = body['ref'].replace('refs/heads/', '')
-    sha = body['head_commit']['id']
-    status_url = body['repository']['statuses_url'].replace('{sha}', sha)
-
-    build = Build.objects.create(
-        repo=repo,
-        ref=branch,
-        sha=sha,
-        status=Build.PENDING
-    )
-
-    def publish_status(state, description, target_url=None):
-        data = {
-            'state': state,
-            'description': description,
-            'target_url': target_url,
-            'context': 'linty'
-        }
-        requests.post(status_url, json=data, auth=auth)
-
-    publish_status('pending', 'Linting your code...')
-
-    # download repo
-    if not os.path.exists('tmp'):
-        os.makedirs('tmp')
-    directory = 'tmp/%s' % sha[:7]
-    if os.path.exists(directory):
-        shutil.rmtree(directory)
-    subprocess.call(['git', 'clone', clone_url, directory])
-    subprocess.call(['git', '--git-dir=%s/.git' % directory, '--work-tree=%s' % directory, 'fetch', clone_url])
-    subprocess.call(['git', '--git-dir=%s/.git' % directory, '--work-tree=%s' % directory, 'checkout', branch])
-
-    # run linting
-    output = None
-    try:
-        subprocess.check_output(['pep8', directory])
-    except subprocess.CalledProcessError as e:
-        # pep8 returns a non-zero code when it finds issues, so we have to catch the error to get the output
-        output = e.output
-
-    # nuke files
-    shutil.rmtree(directory)
-
-    # process output
-    if not output:
-        status = 'success'
-        publish_status(status, 'Your code passed linting.')
-    else:
-        status = 'error'
-        output = output.replace(directory, '')
-        path = reverse('build_detail', kwargs={'pk': build.id})
-        url = request.build_absolute_uri(path)
-        publish_status(status, 'Your code has lint failures. See Details.', target_url=url)
-
-    # update build record
-    build.status = status
-    build.result = output
-    build.finished_at = timezone.now()
-    build.save()
-
-    return HttpResponse(status=204)
+    return HttpResponse(status=202)
 
 
 class BadgeView(generic.DetailView):
