+++ linty/urls.py
@@ -5,6 +5,7 @@ urlpatterns = [
     url('', include('interface.urls')),
     url('', include('social.apps.django_app.urls', namespace='social')),
     url(r'^admin/', admin.site.urls),
+     url(r'^worker/', include('django_rq.urls')),
 ]
 
 handler404 = 'interface.views.handler404'
