+++ requirements.txt
@@ -7,3 +7,4 @@ python-social-auth==0.2.19
 gunicorn==19.6.0
 whitenoise==3.1
 pygithub==1.26.0
+django-rq==0.9.1
